requirements.txt
ib_insync>=0.9.86
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
scipy>=1.11
backtrader>=1.9.78.123
matplotlib>=3.8
plotly>=5.17
textblob>=0.17
tweepy>=4.14
httpx>=0.25
pyyaml>=6.0
loguru>=0.7

# Enhanced technical analysis
pandas-ta>=0.3.14b
TA-Lib>=0.4.28

# ML-based optimization
optuna>=3.5.0

# Dashboard backend
fastapi>=0.120.0
uvicorn>=0.38.0
websockets>=15.0

# Reinforcement learning (optional)
# stable-baselines3>=2.2.1
# gymnasium>=0.29.1

# Additional utilities
python-dateutil>=2.8.2
joblib>=1.3.2
